pyproject.toml
[tool.poetry]
name = "advanced-python-course"
version = "0.1.0"
description = "Advanced Python Course"
authors = ["Simon Strandman <you@example.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
markdown = "^3.6"
pytest = "^8.4.1"
requests = "^2.32.4"
types-requests = "^2.32.4.20250809"
mypy = "^1.17.1"
types-markdown = "^3.9.0.20250906"
pygments = "^2.19.2"
pymdown-extensions = "^10.16.1"
flask = "^3.1.2"
pydantic = "^2.11.9"
pyright = "^1.1.405"
ruff = "^0.13.1"
google-genai = "^1.38.0"
dotenv = "^0.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"